requirements.txt
asgiref==3.5.2
distlib==0.3.4
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
filelock==3.7.1
joblib==1.1.0
mysqlclient==2.1.1
platformdirs==2.5.2
psycopg2==2.9.3
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tzdata==2022.1
virtualenv==20.15.0